Implemented a saner behaviour for moving frame(s). We consider the bounding rect of all the frames we are moving, and check that it's contained within the page. This fixes table cells being smashed up against a border, and allows to get rid of oldMx/oldMy. I couldn't get the QCursor::setPos() to work as it should, though - anyone ?
svn path=/trunk/koffice/; revision=92708

+++ kword/kwdoc.cc
@@ -2338,9 +2338,9 @@ int KWDocument::getPageOfRect( QRect & _rect ) const
 bool KWDocument::isOutOfPage( QRect & r, int page ) const
 {
     return r.x() < 0 ||
-        r.right() > static_cast<int>( ptPaperWidth() ) ||
-        r.y() < page * static_cast<int>( ptPaperHeight() ) ||
-        r.bottom() > ( page + 1 ) * static_cast<int>( ptPaperHeight() );
+        r.right() > ptPaperWidth() ||
+        r.y() < page * ptPaperHeight() ||
+        r.bottom() > ( page + 1 ) * ptPaperHeight();
 }
 
 bool KWDocument::selection() {
